Refactor rendering of project header
